antimony::cli::run

Struct Args

pub struct Args {

    pub profile: String,
    pub dry: bool,
    pub log: bool,
    pub refresh: bool,
    pub path: Option<String>,
    pub config: Option<String>,
    pub features: Option<Vec<String>>,
    pub conflicts: Option<Vec<String>>,
    pub inherits: Option<Vec<String>>,
    pub home_policy: Option<HomePolicy>,
    pub home_name: Option<String>,
    pub home_path: Option<String>,
    pub home_lock: Option<bool>,
    pub seccomp: Option<SeccompPolicy>,
    pub portals: Option<Vec<Portal>>,
    pub see: Option<Vec<String>>,
    pub talk: Option<Vec<String>>,
    pub own: Option<Vec<String>>,
    pub call: Option<Vec<String>>,
    pub disable_ipc: bool,
    pub system_bus: bool,
    pub user_bus: bool,
    pub file_passthrough: Option<FileMode>,
    pub ro: Option<Vec<String>>,
    pub rw: Option<Vec<String>>,
    pub binaries: Option<Vec<String>>,
    pub libraries: Option<Vec<String>>,
    pub devices: Option<Vec<String>>,
    pub namespaces: Option<Vec<Namespace>>,
    pub env: Option<Vec<String>>,
    pub new_privileges: Option<bool>,
    pub sandbox_args: Option<Vec<String>>,
    pub passthrough: Option<Vec<String>>,

}

Fields
profile: String

The name of the profile, or a command to sandbox.

If no profile with the desired name exists, it will be assumed as an arbitrary binary that will be looked up in PATH. It will then be run sandboxed, to which all aspects of the profile can be configured via the command line.

This allows you to run any application in Antimony, without needing to make a profile for it (Although it’s strongly recommended to create a profile for integration and repeated use).

dry: bool

Generate the profile, but do not run the executable.

log: bool

Collect output from the sandbox, and output it to a log file.

refresh: bool

Refresh cache definitions. Analogous to antimony refresh

path: Option<String>

The path to the binary

config: Option<String>

Use a configuration within the profile.

features: Option<Vec<String>>

Additional features.

conflicts: Option<Vec<String>>

Conflicting features

inherits: Option<Vec<String>>

Additional inheritance.

home_policy: Option<HomePolicy>

Override the home policy

home_name: Option<String>

Override the home name

home_path: Option<String>

Override the home mount

home_lock: Option<bool>

Override the home lock

seccomp: Option<SeccompPolicy>

Override the seccomp policy

portals: Option<Vec<Portal>>

Add portals

see: Option<Vec<String>>

Add busses the sandbox can see.

talk: Option<Vec<String>>

Add busses the sandbox can talk to.

own: Option<Vec<String>>

Add busses the sandbox owns.

call: Option<Vec<String>>

Add busses the sandbox can call.

disable_ipc: bool

Disable all IPC. This overrules all other IPC settings.

system_bus: bool

Provide the system bus.

user_bus: bool

Provide the user bus. xdg-dbus-proxy is not run.

file_passthrough: Option<FileMode>

Override the file passthrough mode.

ro: Option<Vec<String>>

Add read-only files

rw: Option<Vec<String>>

Add read-write files.

binaries: Option<Vec<String>>

Add binaries

libraries: Option<Vec<String>>

Add libraries

devices: Option<Vec<String>>

Add devices

namespaces: Option<Vec<Namespace>>

Add namespaces

env: Option<Vec<String>>

Add environment variables in KEY=VALUE syntax

new_privileges: Option<bool>sandbox_args: Option<Vec<String>>

Arguments to pass to bubblewrap/wrapper

passthrough: Option<Vec<String>>

Arguments to pass to the profile application.