antimony

Module fab

Shared functionality between the Fabricators. This contains the brunt of the library fabricator’s logic.

Modules

bin

The Binary Fabricator determines all executables that are used by the program by analyzing it underneath a specialized SECCOMP Notifier.

dev

features

The Feature Fabricator composes all defined features defined in the Profile, recursively analyzes dependencies, strikes conflicts, and merges definitions into a single, complete Profile.

files

lib

The Library Fabricator is most important of all, as it assembles the SOF.

ns

Structs

FabInfo

Each fabricator is passed this structure.

Statics

ELF_MAGIC

The magic for an ELF file.

Functions

get_dir

Get all executable files in a directory. This is very expensive.

get_libraries

LDD a path.

get_wildcards

Find all matches in a directory. We only match the top level for performance considerations.

in_lib

ldd

LDD an executable to get its library dependencies.

localize_home

Localize a home path to /home/antimony

localize_path

Localize a path, returning the resolved source, and where it should be mounted in the sandbox. This is a rather complicated function because it handles a lot of different cases. Specifically:

resolve

Resolve environment variables within paths.