antimony::fab::bin

Struct ParseReturn

pub struct ParseReturn {
    pub elf: Set<String>,
    pub files: Set<String>,
    pub scripts: Set<String>,
    pub symlinks: Set<(String, String)>,
    pub localized: Map<String, String>,
    pub directories: Set<String>,
}

Information returned from parse.

Fields
elf: Set<String>

ELF files, to be passed to the library fabricator.

files: Set<String>

Regular files, which act as heuristics for library folders.

scripts: Set<String>

Script files, which need no further parsing, but must be mounted.

symlinks: Set<(String, String)>

Symlinks

localized: Map<String, String>

Localized values

directories: Set<String>

Library directories.