antimony::fab

Function localize_path

pub fn localize_path(
    file: &str,
    home: bool,
) -> Result<(Option<Cow<'_, str>>, String), Error>

Localize a path, returning the resolved source, and where it should be mounted in the sandbox. This is a rather complicated function because it handles a lot of different cases. Specifically:

Resolves environment variables ($AT_HOME => /usr/share/antimony)

Resolves home markers => (~/test => $HOME/test)

The destination is mapped to Antimony’s home (/home/test/file => /home/antimony/file)

When the home flag is set, paths can be relative to home (file => $HOME/file)

An explicit mapping can be provided with an equal sign. ($AT_HOME/binary=/usr/bin/binary => /usr/bin/binary)

The return value is also rather complicated, and not entirely intuitive. The src aspect is optional, but the destination is always returned. That’s because the source is the important aspect (Its the location on the host), so if it doesn’t exist, the callers can just check src and drop the entire thing otherwise.

Examples

use antimony::fab::localize_path;
use std::borrow::Cow;

unsafe {
    std::env::set_var("HOME", "/home/test");
    std::env::set_var("UID", "1000");
    std::env::set_var("PID", "1");
    std::env::set_var("FD", "2");
}

let temp = temp::Builder::new().create::<temp::File>().unwrap();
let path = temp.full();
let str = path.to_string_lossy();

assert!(localize_path(str.clone().as_ref(), false).unwrap() == (Some(str.clone()), str.into_owned()));

assert!(localize_path("/etc/file", false).unwrap() == (None, "/etc/file".to_owned()));
assert!(localize_path("/etc/file=/file", false).unwrap() == (None, "/file".to_owned()));

assert!(localize_path("file", true).unwrap() == (None, "/home/antimony/file".to_owned()));
assert!(localize_path("~/file", true).unwrap() == (None, "/home/antimony/file".to_owned()));
assert!(localize_path("$HOME/file", true).unwrap() == (None, "/home/antimony/file".to_owned()));

assert!(localize_path("/run/$UID/file", false).unwrap() == (None, "/run/1000/file".to_owned()));
assert!(localize_path("/proc/$PID/fd/$FD=/init-err", false).unwrap() == (None, "/init-err".to_owned()));