antimony

Module shared

Modules

config

The configuration file is a global configuration for Antimony.

edit

Edit a file.

env

Environment Variables Antimony needs defined.

feature

Features are miniature profiles used by the latter for common functionality.

profile

The Profile defines what application should be run, and what features it needs to function properly. It’s the core of Antimony, and has been separated into separate files for readability.

syscalls

The bulk of the SECCOMP Logic.

Macros

timer

Debug macro to record how long something took, but only in developer builds. On release builds, this does nothing.

Structs

StaticHash

Functions

direct_path

Get where direct files should be placed.

format_iter

Format an iterator into a string.

logger

Our notify logger implementation. Because Antimony runs SetUID, we have to spawn a separate process to access the user bus.

privileged

Check that the Real User is privileged. This is used to allow modifying the Antimony system, it does not correlate to actual administrative access (IE sudo/polkit)

user_dir

The user dir is where the instance information is stored.

utility

Get the path to a utility.

Type Aliases

Map

Set