antimony::shared

Module profile

The Profile defines what application should be run, and what features it needs to function properly.

Structs

Files

Files, RO/RW, and Modes.

Home

Hook

A Hook is a program run in coordination with the profile. Hooks are run as the user. Hooks are invoked with the following environment variables: ANTIMONY_NAME: The name of the current profile. ANTIMONY_HOME: The path to the home folder, if it exists. ANTIMONY_CACHE: The cache of the profile in /usr/share/antimony/cache

Hooks

The Hooks structure contains both pre and post hooks.

Ipc

IPC mediated via xdg-dbus-proxy.

Profile

The definitions needed to sandbox an application.

Enums

Error

An error for issues around Profiles.

FileMode

HomePolicy

The Home Policy being set creates a persistent home folder for the profile.

HookError

An error for hooks.

Namespace

Namespaces. By default, none are shared. You will likely not need to use these directly, as they are included in relevant features.

Portal

A non-exhaustive list of Portals. Some may not be implemented for certain Desktop Environments. Not all applications use portals, even if they are provided to the sandbox.

SeccompPolicy

The SECCOMP Policy for the Profile

Statics

CACHE_DIR

FILE_MODES

Functions

library_info

Print info about the libraries used in a feature/profile.

Type Aliases

FileList

Files can either be passed Read-Only, or ReadWrite. Note that some applications, particularly KDE, do not write to the file directly; they make a copy in the same directory (In this case the sandbox), then move it from source to destination to replace it. However, because Antimony provides file as bind mounts, this operation will fail. In such cases, use portals, put the file directly in the profile’s home folder, or pass the parent folder.