antimony::shared::profile::files

Struct Files

pub struct Files {
    pub passthrough: Option<FileMode>,
    pub temp: Vec<String>,
    pub user: FileList,
    pub platform: FileList,
    pub resources: FileList,
    pub direct: Map<FileMode, Map<String, String>>,
}

Files, RO/RW, and Modes.

Fields
passthrough: Option<FileMode>

The default mode for files passed through the command line. If no passthrough is provided, files are not passed. This includes using the application to open files in your file explorer or setting it as the default for particular MIME types.

temp: Vec<String>

A list of paths in the sandbox to mount as TMPFS.

user: FileList

User files assume a root of /home/antimony unless absolute.

platform: FileList

Platform files are device-specific system files (Locale, Configuration, etc)

resources: FileList

Resource files are system files required by libraries/binaries.

direct: Map<FileMode, Map<String, String>>

Direct files take a path, and file contents.

Implementations

impl Files

pub fn merge(&mut self, files: Self)

Merge two file sets together.

pub fn from_args(args: &mut Args) -> Option<Self>

Construct a file set from the command line.