antimony::shared::profile::hooks

Struct Hook

pub struct Hook {
    pub name: Option<String>,
    pub t: Type,
    pub content: String,
    pub arguments: Vec<String>,
    pub attach: Option<bool>,
    pub env: Option<bool>,
    pub can_fail: Option<bool>,
    pub new_privileges: Option<bool>,
    pub capture_output: Option<bool>,
    pub capture_error: Option<bool>,
}

A Hook is a program run in coordination with the profile.

Hooks are run as the user. Hooks are invoked with the following environment variables: ANTIMONY_NAME: The name of the current profile. ANTIMONY_HOME: The path to the home folder, if it exists. ANTIMONY_CACHE: The cache of the profile in /usr/share/antimony/cache

Fields
name: Option<String>

An optional name to identify the process.

t: Typecontent: String

The contents of the hook. For scripts, the full content. For programs, the name/path of the binary. For antimony, the profile

arguments: Vec<String>

A list of arguments to be passed to the hook

attach: Option<bool>

In pre-hooks a hook can be attached to the sandbox. In this mode, the hook runs alongside the sandbox. If false, the program waits for the hook to terminate before launching the sandbox.

env: Option<bool>

Share the environment with the hook.

can_fail: Option<bool>

If the Hook can fail. If false, an error will abort the program.

new_privileges: Option<bool>

Allow the hook to obtain new privileges. Needed if the binary/script requires privilege antimony does not have.

capture_output: Option<bool>

Capture the sandbox’s STDOUT and provide it to the Hook via STDIN. Only one of capture_output and capture_error can be set. If both are set, capture_error is used.

capture_error: Option<bool>

Capture the sandbox’s STDERR and provide it to the Hook via STDIN. Only one of capture_output and capture_error can be set. If both are set, capture_error is used.

Implementations

impl Hook

pub fn process( &mut self, main: Option<Spawner>, name: &str, cache: &str, instance: &str, home: &Option<String>, parent: bool, ) -> Result<Option<Spawner>, HookError>

Process the hook.