antimony::shared::profile::ipc

Enum Portal

pub enum Portal {

    Background = 0,
    Camera = 1,
    Clipboard = 2,
    Documents = 3,
    FileChooser = 4,
    GlobalShortcuts = 5,
    Inhibit = 6,
    Location = 7,
    Notifications = 8,
    OpenURI = 9,
    ProxyResolver = 10,
    Realtime = 11,
    ScreenCast = 12,
    Screenshot = 13,
    Settings = 14,
    Secret = 15,
    NetworkMonitor = 16,

}

A non-exhaustive list of Portals. Some may not be implemented for certain Desktop Environments. Not all applications use portals, even if they are provided to the sandbox.

Variants

Background = 0

Camera = 1

Clipboard = 2

Documents = 3

FileChooser = 4

GlobalShortcuts = 5

Inhibit = 6

Location = 7

Notifications = 8

OpenURI = 9

ProxyResolver = 10

Realtime = 11

ScreenCast = 12

Screenshot = 13

Settings = 14

Secret = 15

NetworkMonitor = 16