antimony

Module fab

Shared functionality between the Fabricators. This contains the brunt of the library fabricator’s logic.

Modules

bin

The Binary Fabricator determines all executables that are used by the program by analyzing it underneath a specialized SECCOMP Notifier. It then passes these binaries to the Library Fabricator to analyze under LDD.

dev

etc

features

The Feature Fabricator composes all defined features defined in the Profile, recursively analyzes dependencies, strikes conflicts, and merges definitions into a single, complete Profile.

files

lib

The Library Fabricator is most important of all, as it assembles the SOF. It also touches almost every other fabricator, and is the central part of the most important path between bin-library-syscalls. It is also by the far the most complicated, as libraries can encompass files, wildcards, directories, and binaries. They can be sourced from just about anywhere on the system (IE /usr/bin, or /usr/share/application), and it needs to determine which files should be placed in the SOF, and what to do with their dependencies. It relies on LDD to determine ELF dependencies (IE .so files), and Find to scour directories. Everything is aggressively cached, and even more aggressively parallelized.

ns

Structs

FabInfo

Each fabrciator is passed this structure.

Statics

ELF_MAGIC

The magic for an ELF file.

LIB_ROOTS

Get the library roots.

Functions

get_dir

Get all executable files in a directory. This is very expensive.

get_libraries

LDD a path.

get_wildcards

Find all matches in a directory. We only match the top level for performance considerations.

localize_home

Localize a home path to /home/antimony

localize_path

Ensure ~ points to /home/antimony

resolve

Resolve environment variables in paths.

resolve_env

Resolve environment variables within names.